detection/requirements.txt
###### to install all needed libraries: pip install -r requirements.txt ######
# Use pip show <package name> to see version
ultralytics>=8.0.121
tk>=0.1.0
pillow>=9.5.0